requirements.txt
asgiref==3.3.1
Django==3.1.4
django-js-asset==1.2.2
django-mptt==0.11.0
pytz==2020.4
sqlparse==0.4.1
django-crispy-forms==1.9.2
Pillow==7.2.0

# For herokuapp
gunicorn==20.0.4
whitenoise==5.1.0
psycopg2==2.8.5
